+++ Include/Acidanthera/Library/OcConfigurationLib.h
@@ -150,6 +150,7 @@
   _(BOOLEAN                     , ProvideCustomSlide        ,     , FALSE  , ()) \
   _(UINT8                       , ProvideMaxSlide           ,     , 0      , ()) \
   _(BOOLEAN                     , RebuildAppleMemoryMap     ,     , FALSE  , ()) \
+  _(INT8                        , ResizeAppleGpuBars        ,     , -1     , ()) \
   _(BOOLEAN                     , SetupVirtualMap           ,     , FALSE  , ()) \
   _(BOOLEAN                     , SignalAppleOS             ,     , FALSE  , ()) \
   _(BOOLEAN                     , SyncRuntimePermissions    ,     , FALSE  , ())
@@ -572,9 +573,10 @@ typedef enum {
 /// Drivers is an ordered array of drivers to load.
 ///
 #define OC_UEFI_DRIVER_ENTRY_FIELDS(_, __) \
-  _(OC_STRING                   , Path               ,     , OC_STRING_CONSTR ("", _, __)                    , OC_DESTR (OC_STRING) ) \
+  _(OC_STRING                   , Arguments          ,     , OC_STRING_CONSTR ("", _, __)                    , OC_DESTR (OC_STRING) ) \
+  _(OC_STRING                   , Comment            ,     , OC_STRING_CONSTR ("", _, __)                    , OC_DESTR (OC_STRING) ) \
   _(BOOLEAN                     , Enabled            ,     , FALSE                                           , ())                    \
-  _(OC_STRING                   , Arguments          ,     , OC_STRING_CONSTR ("", _, __)                    , OC_DESTR (OC_STRING) )
+  _(OC_STRING                   , Path               ,     , OC_STRING_CONSTR ("", _, __)                    , OC_DESTR (OC_STRING) )
   OC_DECLARE (OC_UEFI_DRIVER_ENTRY)
 
 #define OC_UEFI_DRIVER_ARRAY_FIELDS(_, __) \
@@ -689,6 +691,7 @@ typedef enum {
   _(BOOLEAN                     , EnableVectorAcceleration    ,     , FALSE  , ()) \
   _(BOOLEAN                     , ForgeUefiSupport            ,     , FALSE  , ()) \
   _(BOOLEAN                     , IgnoreInvalidFlexRatio      ,     , FALSE  , ()) \
+  _(INT8                        , ResizeGpuBars               ,     , -1     , ()) \
   _(BOOLEAN                     , ReleaseUsbOwnership         ,     , FALSE  , ()) \
   _(BOOLEAN                     , ReloadOptionRoms            ,     , FALSE  , ()) \
   _(BOOLEAN                     , RequestBootVarRouting       ,     , FALSE  , ()) \