+++ Include/Acidanthera/Library/OcMainLib.h
@@ -31,7 +31,7 @@
   OpenCore version reported to log and NVRAM.
   OPEN_CORE_VERSION must follow X.Y.Z format, where X.Y.Z are single digits.
 **/
-#define OPEN_CORE_VERSION          "0.7.3"
+#define OPEN_CORE_VERSION          "0.7.5"
 
 /**
   OpenCore build type reported to log and NVRAM.
@@ -189,6 +189,17 @@ VOID
 OcGetLegacySecureBootECID (
   IN  OC_GLOBAL_CONFIG    *Config,
   OUT UINT64              *ApECID
+  );
+
+/**
+  Obtain default Secure Boot hardware model.
+
+  @param[in]  Config   Configuration.
+  @return Apple Secure Boot model.
+**/
+CONST CHAR8 *
+OcGetDefaultSecureBootModel (
+  IN  OC_GLOBAL_CONFIG    *Config
   );
 
 /**