+++ Library/OcMainLib/OpenCorePlatform.c
@@ -371,10 +371,12 @@ OcPlatformUpdateSmbios (
     if (Config->PlatformInfo.Generic.AdviseFeatures) {
       Data.FirmwareFeatures     |= FW_FEATURE_SUPPORTS_CSM_LEGACY_MODE
         | FW_FEATURE_SUPPORTS_UEFI_WINDOWS_BOOT
-        | FW_FEATURE_SUPPORTS_APFS;
+        | FW_FEATURE_SUPPORTS_APFS
+        | FW_FEATURE_SUPPORTS_LARGE_BASESYSTEM;
       Data.FirmwareFeaturesMask |= FW_FEATURE_SUPPORTS_CSM_LEGACY_MODE
         | FW_FEATURE_SUPPORTS_UEFI_WINDOWS_BOOT
-        | FW_FEATURE_SUPPORTS_APFS;
+        | FW_FEATURE_SUPPORTS_APFS
+        | FW_FEATURE_SUPPORTS_LARGE_BASESYSTEM;
     }
 
     //
@@ -538,10 +540,12 @@ OcPlatformUpdateNvram (
     if (Config->PlatformInfo.Generic.AdviseFeatures) {
       ExFeatures     |= FW_FEATURE_SUPPORTS_CSM_LEGACY_MODE
         | FW_FEATURE_SUPPORTS_UEFI_WINDOWS_BOOT
-        | FW_FEATURE_SUPPORTS_APFS;
+        | FW_FEATURE_SUPPORTS_APFS
+        | FW_FEATURE_SUPPORTS_LARGE_BASESYSTEM;
       ExFeaturesMask |= FW_FEATURE_SUPPORTS_CSM_LEGACY_MODE
         | FW_FEATURE_SUPPORTS_UEFI_WINDOWS_BOOT
-        | FW_FEATURE_SUPPORTS_APFS;
+        | FW_FEATURE_SUPPORTS_APFS
+        | FW_FEATURE_SUPPORTS_LARGE_BASESYSTEM;
     }
   }
 
@@ -929,6 +933,60 @@ OcGetLegacySecureBootECID (
   }
 
   CopyMem (ApECID, &Uuid, sizeof (*ApECID));
+}
+
+CONST CHAR8 *
+OcGetDefaultSecureBootModel (
+  IN  OC_GLOBAL_CONFIG    *Config
+  )
+{
+  EFI_STATUS        Status;
+  CONST CHAR8       *Model;
+  CONST CHAR8       *Board;
+  CONST CHAR8       *SbModel;
+  OC_SMBIOS_TABLE   SmbiosTable;
+
+  //
+  // For automatic setups it is direct DB retrieval.
+  //
+  if (Config->PlatformInfo.Automatic) {
+    Model   = OC_BLOB_GET (&Config->PlatformInfo.Generic.SystemProductName);
+    SbModel = GetSecureBootModel (Model);
+    DEBUG ((DEBUG_INFO, "OC: Automatic SB model %a from model %a\n", SbModel, Model));
+    return SbModel;
+  }
+
+  //
+  // For manual setups only use the SMBIOS board identifier.
+  //
+  Board = OC_BLOB_GET (&Config->PlatformInfo.Smbios.BoardProduct);
+  if (Config->PlatformInfo.UpdateSmbios && Board[0] != '\0') {
+    SbModel = GetSecureBootModelFromBoardId (Board);
+    DEBUG ((DEBUG_INFO, "OC: Manual SB model %a from board %a\n", SbModel, Board));
+    return SbModel;
+  }
+
+  //
+  // For Mac setups without spoofing use SMBIOS.
+  //
+  Status = OcSmbiosTablePrepare (&SmbiosTable);
+  if (!EFI_ERROR (Status)) {
+    OcSmbiosExtractOemInfo (
+      &SmbiosTable,
+      mCurrentSmbiosProductName,
+      NULL,
+      NULL,
+      NULL,
+      NULL,
+      Config->PlatformInfo.UseRawUuidEncoding,
+      FALSE
+      );
+    OcSmbiosTableFree (&SmbiosTable);
+  }
+
+  SbModel = GetSecureBootModel (mCurrentSmbiosProductName);
+  DEBUG ((DEBUG_INFO, "OC: OEM SB model %a from model %a\n", SbModel, mCurrentSmbiosProductName));
+  return SbModel;
 }
 
 BOOLEAN