+++ Platform/OpenLinuxBoot/LoaderEntry.c
@@ -23,11 +23,19 @@
 #include <Protocol/OcBootEntry.h>
 
 //
+// Root.
+//
+#define ROOT_DIR    L"\\"
+//
 // Required where the BTRFS subvolume is /boot, as this looks like a
 // normal directory within EFI. Note that scanning / and then /boot
 // is how blscfg behaves by default too.
 //
-#define ADDITIONAL_SCAN_DIR   L"boot"
+#define BOOT_DIR    L"\\boot"
+//
+// Don't add missing root= option unless this directory is present at root.
+//
+#define OSTREE_DIR  L"\\ostree"
 
 //
 // No leading slash so they can be relative to root or
@@ -249,19 +257,21 @@ InternalAllocateLoaderEntry (
 {
   LOADER_ENTRY      *Entry;
 
-  Entry = AllocateZeroPool (sizeof (LOADER_ENTRY));
+  Entry = OcFlexArrayAddItem (gLoaderEntries);
 
   if (Entry != NULL) {
     Entry->Options = OcFlexArrayInit (sizeof (CHAR8 *), OcFlexArrayFreePointerItem);
     if (Entry->Options == NULL) {
-      InternalFreeLoaderEntry (&Entry);
+      InternalFreeLoaderEntry (Entry);
+      Entry = NULL;
     }
   }
 
   if (Entry != NULL) {
     Entry->Initrds = OcFlexArrayInit (sizeof (CHAR8 *), OcFlexArrayFreePointerItem);
     if (Entry->Initrds == NULL) {
-      InternalFreeLoaderEntry (&Entry);
+      InternalFreeLoaderEntry (Entry);
+      Entry = NULL;
     }
   }
 
@@ -270,34 +280,31 @@ InternalAllocateLoaderEntry (
 
 VOID
 InternalFreeLoaderEntry (
-  LOADER_ENTRY    **Entry
+  LOADER_ENTRY    *Entry
   )
 {
   ASSERT (Entry != NULL);
-  ASSERT (*Entry != NULL);
 
-  if (Entry != NULL && *Entry != NULL) {
-    if ((*Entry)->Title) {
-      FreePool ((*Entry)->Title);
-    }
-    if ((*Entry)->Version) {
-      FreePool ((*Entry)->Version);
-    }
-    if ((*Entry)->Linux) {
-      FreePool ((*Entry)->Linux);
-    }
-    OcFlexArrayFree (&(*Entry)->Initrds);
-    OcFlexArrayFree (&(*Entry)->Options);
-    if ((*Entry)->OcId != NULL) {
-      FreePool ((*Entry)->OcId);
-    }
-    if ((*Entry)->OcFlavour != NULL) {
-      FreePool ((*Entry)->OcFlavour);
-    }
+  if (Entry == NULL) {
+    return;
+  }
 
-    FreePool (*Entry);
-
-    *Entry = NULL;
+  if (Entry->Title) {
+    FreePool (Entry->Title);
+  }
+  if (Entry->Version) {
+    FreePool (Entry->Version);
+  }
+  if (Entry->Linux) {
+    FreePool (Entry->Linux);
+  }
+  OcFlexArrayFree (&Entry->Initrds);
+  OcFlexArrayFree (&Entry->Options);
+  if (Entry->OcId != NULL) {
+    FreePool (Entry->OcId);
+  }
+  if (Entry->OcFlavour != NULL) {
+    FreePool (Entry->OcFlavour);
   }
 }
 
@@ -352,7 +359,7 @@ EFI_STATUS
 InternalProcessLoaderEntryFile (
   IN     CONST CHAR16           *FileName,
   IN OUT       CHAR8            *Content,
-     OUT       LOADER_ENTRY     **Entry,
+     OUT       LOADER_ENTRY     *Entry,
   IN     CONST BOOLEAN          Grub2
   )
 {
@@ -362,11 +369,6 @@ InternalProcessLoaderEntryFile (
   CHAR8         *Value;
 
   ASSERT (Content != NULL);
-
-  *Entry = InternalAllocateLoaderEntry ();
-  if (*Entry == NULL) {
-    return EFI_OUT_OF_RESOURCES;
-  }
 
   Status = EFI_SUCCESS;
   Pos = 0;
@@ -386,24 +388,22 @@ InternalProcessLoaderEntryFile (
     }
 
     if (EFI_ERROR (Status)) {
-      InternalFreeLoaderEntry (Entry);
       return Status;
     }
 
     if (AsciiStrCmp (Key, "title") == 0) {
-      Status = EntryCopySingleValue (Grub2, &(*Entry)->Title, Value);
+      Status = EntryCopySingleValue (Grub2, &Entry->Title, Value);
     } else if (AsciiStrCmp (Key, "version") == 0) {
-      Status = EntryCopySingleValue (Grub2, &(*Entry)->Version, Value);
+      Status = EntryCopySingleValue (Grub2, &Entry->Version, Value);
     } else if (AsciiStrCmp (Key, "linux") == 0) {
-      Status = EntryCopySingleValue (Grub2, &(*Entry)->Linux, Value);
+      Status = EntryCopySingleValue (Grub2, &Entry->Linux, Value);
     } else if (AsciiStrCmp (Key, "options") == 0) {
-      Status = EntryCopyMultipleValue (Grub2, (*Entry)->Options, Value);
+      Status = EntryCopyMultipleValue (Grub2, Entry->Options, Value);
     } else if (AsciiStrCmp (Key, "initrd") == 0) {
-      Status = EntryCopyMultipleValue (FALSE, (*Entry)->Initrds, Value);
+      Status = EntryCopyMultipleValue (FALSE, Entry->Initrds, Value);
     }
 
     if (EFI_ERROR (Status)) {
-      InternalFreeLoaderEntry (Entry);
       return Status;
     }
   }
@@ -503,77 +503,225 @@ DoFilterLoaderEntry (
 // If autodetect, we assume there is only one install in the directory,
 // so {vmlinuz} is good as the shared id.
 //
-STATIC
-CHAR8 *
-OcIdFromFileName (
-  CHAR16      *FileName
+EFI_STATUS
+InternalIdVersionFromFileName (
+  IN OUT LOADER_ENTRY  *Entry,
+  IN     CHAR16        *FileName
   )
 {
   EFI_STATUS          Status;
   UINTN               NumCopied;
 
+  CHAR16              *Split;
   CHAR16              *IdEnd;
+  CHAR16              *VersionStart;
+  CHAR16              *VersionEnd;
   CHAR8               *OcId;
+  CHAR8               *Version;
 
+  ASSERT (Entry    != NULL);
   ASSERT (FileName != NULL);
   
-  IdEnd   = NULL;
+  Split = NULL;
 
-  //
-  // Shared id; intended to pick up machine-id from standard .conf file
-  // naming, or just the word vmlinuz from standard kernel file naming.
-  //
-  if ((gLinuxBootFlags & LINUX_BOOT_USE_LATEST) != 0) {
-    IdEnd = OcStrChr (FileName, L'-');
+  Split = OcStrChr (FileName, L'-');
+
+  VersionEnd = &FileName[StrLen (FileName)];
+  if (OcUnicodeEndsWith (FileName, BLSPEC_SUFFIX_CONF, TRUE)) {
+    VersionEnd -= L_STR_LEN (BLSPEC_SUFFIX_CONF);
   }
 
   //
-  // Non-shared id, or no '-' found in filename above.
+  // Will not happen with sane filenames.
   //
-  if (IdEnd == NULL) {
-    IdEnd = &FileName[StrLen (FileName)];
-    if (OcUnicodeEndsWith (FileName, L".conf", TRUE)) {
-      IdEnd -= L_STR_LEN (L".conf");
+  if (Split != NULL && (Split == FileName || VersionEnd == Split + 1)) {
+    Split = NULL;
+  }
+
+  if (Split == NULL) {
+    VersionStart = FileName;
+    IdEnd = VersionEnd;
+  } else {
+    VersionStart = Split + 1;
+
+    if ((gLinuxBootFlags & LINUX_BOOT_USE_LATEST) != 0) {
+      IdEnd = Split;
+    } else {
+      IdEnd = VersionEnd;
     }
   }
 
   OcId = AllocatePool (IdEnd - FileName + 1);
-
-  if (OcId != NULL) {
-    Status = UnicodeStrnToAsciiStrS (FileName, IdEnd - FileName, OcId, IdEnd - FileName + 1, &NumCopied);
-    ASSERT (!EFI_ERROR (Status));
-    ASSERT (NumCopied == (UINTN)(IdEnd - FileName));
-  }
-
-  return OcId;
-}
-
-NAMED_LOADER_ENTRY *
-InternalCreateNamedLoaderEntry (
-  IN  LOADER_ENTRY        *Entry,
-  IN  CHAR16              *FileName
-  )
-{
-  NAMED_LOADER_ENTRY      *NamedEntry;
-  CHAR8                   *OcId;
-
-  OcId = OcIdFromFileName (FileName);
   if (OcId == NULL) {
-    return NULL;
+    return EFI_OUT_OF_RESOURCES;
   }
 
-  NamedEntry = OcFlexArrayAddItem (gNamedLoaderEntries);
-  if (NamedEntry == NULL) {
-    FreePool (OcId);
-    return NULL;
-  }
+  Status = UnicodeStrnToAsciiStrS (FileName, IdEnd - FileName, OcId, IdEnd - FileName + 1, &NumCopied);
+  ASSERT (!EFI_ERROR (Status));
+  ASSERT (NumCopied == (UINTN) (IdEnd - FileName));
 
   Entry->OcId = OcId;
 
-  NamedEntry->Entry = Entry;
-  NamedEntry->FileName = FileName;
+  if (Entry->Version == NULL) {
+    Version = AllocatePool (VersionEnd - VersionStart + 1);
+    if (Version == NULL) {
+      return EFI_OUT_OF_RESOURCES;
+    }
 
-  return NamedEntry;
+    Status = UnicodeStrnToAsciiStrS (VersionStart, VersionEnd - VersionStart, Version, VersionEnd - VersionStart + 1, &NumCopied);
+    ASSERT (!EFI_ERROR (Status));
+    ASSERT (NumCopied == (UINTN) (VersionEnd - VersionStart));
+
+    Entry->Version = Version;
+  }
+
+  return EFI_SUCCESS;
+}
+
+//
+// We do not need to warn about missing files: OC warns about
+// missing bootable file (vmlinuz) and kernel stops and warns
+// about missing initrd. However, some linuxes (e.g. Endless)
+// make the filepaths specified in an entry relative to /boot,
+// not relative to / as it should be, so we have to check which
+// it is. Only for this reason, we end up warning here if we
+// can't find it at all.
+//
+STATIC
+EFI_STATUS
+FindLoaderFile (
+  IN     CONST EFI_FILE_HANDLE  Directory,
+  IN     CONST CHAR16           *DirName,
+  IN OUT       CHAR8            **FileName
+  )
+{
+  EFI_STATUS        Status;
+  EFI_FILE_PROTOCOL *File;
+  UINTN             FileNameLen;
+  UINTN             DirNameLen;
+  CHAR16            *Path;
+  UINTN             MaxPathSize;
+
+  ASSERT (DirName != NULL);
+  ASSERT (FileName != NULL);
+  ASSERT (*FileName != NULL);
+
+  FileNameLen = AsciiStrLen (*FileName);
+  DirNameLen = StrLen (DirName);
+
+  MaxPathSize = (DirNameLen + FileNameLen + 1) * sizeof (CHAR16);
+
+  Path = AllocatePool (MaxPathSize);
+  if (Path == NULL) {
+    return EFI_OUT_OF_RESOURCES;
+  }
+
+  UnicodeSPrintAsciiFormat (Path, MaxPathSize, "%a", *FileName);
+  UnicodeUefiSlashes (Path);
+
+  Status = OcSafeFileOpen (Directory, &File, Path, EFI_FILE_MODE_READ, 0);
+  if (!EFI_ERROR (Status)) {
+    File->Close (File);
+    FreePool (Path);
+    return Status;
+  }
+
+  if (DirNameLen <= 1) {
+    DEBUG ((DEBUG_WARN, "LNX: %a not found - %r\n", *FileName, Status));
+    FreePool (Path);
+    return Status;
+  }
+
+  UnicodeSPrintAsciiFormat (Path, MaxPathSize, "%s%a", DirName, *FileName);
+  UnicodeUefiSlashes (Path);
+
+  Status = OcSafeFileOpen (Directory, &File, Path, EFI_FILE_MODE_READ, 0);
+  if (!EFI_ERROR (Status)) {
+    //
+    // Found at 'wrong' location - re-use allocated path
+    //
+    File->Close (File);
+    AsciiSPrint ((CHAR8 *)Path, MaxPathSize, "%s%a", DirName, *FileName);
+    AsciiUnixSlashes ((CHAR8 *)Path);
+    FreePool (*FileName);
+    *FileName = (CHAR8 *)Path;
+    return EFI_SUCCESS;
+  }
+
+  DEBUG ((DEBUG_WARN, "LNX: %a not found (searched %s and %s) - %r\n", *FileName, ROOT_DIR, DirName, Status));
+  FreePool (Path);
+  return Status;
+}
+
+STATIC
+BOOLEAN
+HasRootOption (
+  IN           OC_FLEX_ARRAY      *Options
+  )
+{
+  EFI_STATUS            Status;
+  UINTN                 Index;
+  CHAR8                 **Option;
+  CHAR8                 *OptionCopy;
+  OC_FLEX_ARRAY         *ParsedVars;
+  BOOLEAN               HasRoot;
+
+  ASSERT (Options != NULL);
+
+  //
+  // TRUE on errors: do not attempt to add root.
+  //
+  for (Index = 0; Index < Options->Count; Index++) {
+    Option = OcFlexArrayItemAt (Options, Index);
+    ASSERT (Option != NULL);
+    ASSERT (*Option != NULL);
+
+    OptionCopy = AllocateCopyPool (AsciiStrSize (*Option), *Option);
+    if (OptionCopy == NULL) {
+      return TRUE;
+    }
+
+    Status = OcParseVars (OptionCopy, &ParsedVars, FALSE);
+    if (EFI_ERROR (Status)) {
+      DEBUG ((DEBUG_WARN, "LNX: Error parsing Options[%u]=<%a> - %r\n", Index, *Option, Status));
+      FreePool (OptionCopy);
+      return TRUE;
+    }
+
+    HasRoot = OcHasParsedVar (ParsedVars, "root", FALSE);
+
+    FreePool (ParsedVars);
+    FreePool (OptionCopy);
+
+    if (HasRoot) {
+      return TRUE;
+    }
+  }
+
+  return FALSE;
+}
+
+STATIC
+BOOLEAN
+HasOstreeDir (
+  EFI_FILE_HANDLE   Directory
+  )
+{
+  EFI_STATUS            Status;
+  EFI_FILE_PROTOCOL     *OstreeFile;
+
+  Status = OcSafeFileOpen (Directory, &OstreeFile, OSTREE_DIR, EFI_FILE_MODE_READ, 0);
+  if (EFI_ERROR (Status)) {
+    return FALSE;
+  }
+
+  Status = OcEnsureDirectoryFile (OstreeFile, TRUE);
+  if (EFI_ERROR (Status)) {
+    DEBUG ((DEBUG_WARN, "LNX: %s found but not a %a - %r\n", OSTREE_DIR, "directory", Status));
+  }
+  OstreeFile->Close (OstreeFile);
+
+  return !EFI_ERROR(Status);
 }
 
 STATIC
@@ -588,58 +736,101 @@ DoProcessLoaderEntry (
   EFI_STATUS            Status;
   CHAR8                 *Content;
   LOADER_ENTRY          *Entry;
-  NAMED_LOADER_ENTRY    *NamedEntry;
-  CHAR16                *FileName;
+  CHAR16                SaveChar;
+  CHAR16                *DirName;
+  CHAR8                 **Initrd;
+  UINTN                 Index;
 
-  //
-  // Doesn't really matter if we return EFI_SUCCESS or EFI_NOT_FOUND in these initial checks;
-  // anything else will terminate directory processing.
-  //
+  ASSERT (Context != NULL);
+  DirName = Context;
+
   if (FileInfoSize > MAX_LOADER_ENTRY_FILE_INFO_SIZE) {
-    DEBUG ((DEBUG_WARN, "LNX: Entry %a %s overlong...\n", "filename", FileInfo->FileName));
-    return EFI_SUCCESS;
+    SaveChar = FileInfo->FileName[MAX_LOADER_ENTRY_NAME_LEN];
+    FileInfo->FileName[MAX_LOADER_ENTRY_NAME_LEN] = CHAR_NULL;
+    DEBUG ((DEBUG_WARN, "LNX: Entry filename overlong: %s...\n", FileInfo->FileName));
+    FileInfo->FileName[MAX_LOADER_ENTRY_NAME_LEN] = SaveChar;
+    return EFI_NOT_FOUND;
   }
 
   if (FileInfo->FileSize > MAX_LOADER_ENTRY_FILE_SIZE) {
-    DEBUG ((DEBUG_WARN, "LNX: Entry %a %s overlong...\n", "file size", FileInfo->FileName));
-    return EFI_SUCCESS;
+    DEBUG ((DEBUG_WARN, "LNX: Entry file size overlong: %s\n", FileInfo->FileName));
+    return EFI_NOT_FOUND;
   }
 
-  DEBUG ((DEBUG_INFO, "LNX: Parsing %s...\n", FileInfo->FileName));
+  DEBUG (((gLinuxBootFlags & LINUX_BOOT_LOG_VERBOSE) == 0 ? DEBUG_VERBOSE : DEBUG_INFO,
+    "LNX: Reading %s...\n", FileInfo->FileName));
 
   Content = OcReadFileFromDirectory (Directory, FileInfo->FileName, NULL, 0);
   if (Content == NULL) {
     return EFI_OUT_OF_RESOURCES;
   }
 
-  Status = InternalProcessLoaderEntryFile (FileInfo->FileName, Content, &Entry, mIsGrub2);
+  Entry = InternalAllocateLoaderEntry ();
+  if (Entry == NULL) {
+    return EFI_OUT_OF_RESOURCES;
+  }
+
+  Status = InternalProcessLoaderEntryFile (FileInfo->FileName, Content, Entry, mIsGrub2);
   if (EFI_ERROR (Status)) {
+    OcFlexArrayDiscardItem (gLoaderEntries, TRUE);
     return Status;
   }
 
   if (Entry->Linux == NULL) {
-    InternalFreeLoaderEntry (&Entry);
-    DEBUG ((DEBUG_INFO, "LNX: No linux line, ignoring\n"));
-  } else {
-    if (mIsGrub2) {
-      Status = ExpandReplaceOptions (Entry);
+    DEBUG (((gLinuxBootFlags & LINUX_BOOT_LOG_VERBOSE) == 0 ? DEBUG_VERBOSE : DEBUG_INFO,
+      "LNX: No linux line, ignoring\n"));
+    OcFlexArrayDiscardItem (gLoaderEntries, TRUE);
+    return EFI_NOT_FOUND;
+  }
+
+  Status = FindLoaderFile (Directory, DirName, &Entry->Linux);
+  if (!EFI_ERROR (Status)) {
+    for (Index = 0; Index < Entry->Initrds->Count; Index++) {
+      Initrd = OcFlexArrayItemAt (Entry->Initrds, Index);
+      Status = FindLoaderFile (Directory, DirName, Initrd);
       if (EFI_ERROR (Status)) {
-        InternalFreeLoaderEntry (&Entry);
+        break;
+      }
+    }
+  }
+  if (EFI_ERROR (Status)) {
+    OcFlexArrayDiscardItem (gLoaderEntries, TRUE);
+    return Status;
+  }
+
+  if (mIsGrub2) {
+    Status = ExpandReplaceOptions (Entry);
+    if (EFI_ERROR (Status)) {
+      OcFlexArrayDiscardItem (gLoaderEntries, TRUE);
+      return Status;
+    }
+  }
+
+  //
+  // Need to understand other reasons to apply this fix (if any),
+  // so not automatically applying unless we recognise the layout.
+  //
+  if ((gLinuxBootFlags & LINUX_BOOT_ALLOW_CONF_AUTO_ROOT) != 0
+    && !HasRootOption (Entry->Options)) {
+    if (!HasOstreeDir (Directory)) {
+      DEBUG ((DEBUG_WARN, "LNX: Missing root option, %s %afound - %afixing\n", OSTREE_DIR, "not ", "not "));
+    } else {
+      DEBUG ((DEBUG_INFO, "LNX: Missing root option, %s %afound - %afixing\n", OSTREE_DIR, "", ""));
+      Status = InsertRootOption (Entry->Options);
+      if (EFI_ERROR (Status)) {
+        OcFlexArrayDiscardItem (gLoaderEntries, TRUE);
         return Status;
       }
     }
+  }
 
-    FileName =  AllocateCopyPool (StrSize (FileInfo->FileName), FileInfo->FileName);
-    if (FileName == NULL) {
-      return EFI_OUT_OF_RESOURCES;
-    }
-
-    NamedEntry = InternalCreateNamedLoaderEntry (Entry, FileName);
-    if (NamedEntry == NULL) {
-      FreePool (FileName);
-      InternalFreeLoaderEntry (&Entry);
-      return EFI_OUT_OF_RESOURCES;
-    }
+  //
+  // Id, and version if not already set within .conf data, from filename.
+  //
+  Status = InternalIdVersionFromFileName (Entry, FileInfo->FileName);
+  if (EFI_ERROR (Status)) {
+    OcFlexArrayDiscardItem (gLoaderEntries, TRUE);
+    return Status;
   }
 
   return EFI_SUCCESS;
@@ -677,8 +868,9 @@ ProcessLoaderEntry (
 
 STATIC
 EFI_STATUS
-InternalScanLoaderEntries (
+ScanLoaderEntriesAtDirectory (
   IN   EFI_FILE_PROTOCOL        *RootDirectory,
+  IN   CHAR16                   *DirName,
   OUT  OC_PICKER_ENTRY          **Entries,
   OUT  UINTN                    *NumEntries
   )
@@ -697,7 +889,7 @@ InternalScanLoaderEntries (
   Status              = EFI_SUCCESS;
   GrubEnv             = NULL;
   mIsGrub2            = FALSE;
-  gNamedLoaderEntries = NULL;
+  gLoaderEntries      = NULL;
 
   //
   // Only treat as GRUB2 if grub2/grub.cfg exists.
@@ -716,11 +908,13 @@ InternalScanLoaderEntries (
       if (GrubEnv == NULL) {
         DEBUG ((DEBUG_WARN, "LNX: %s not found\n", GRUB2_GRUBENV));
       } else {
-        DEBUG ((DEBUG_INFO, "LNX: Reading %s\n", GRUB2_GRUBENV));
+        DEBUG (((gLinuxBootFlags & LINUX_BOOT_LOG_VERBOSE) == 0 ? DEBUG_VERBOSE : DEBUG_INFO,
+          "LNX: Reading %s\n", GRUB2_GRUBENV));
         Status = InternalProcessGrubEnv (GrubEnv, GRUB2_GRUBENV_SIZE);
       }
       if (!EFI_ERROR (Status)) {
-        DEBUG ((DEBUG_INFO, "LNX: Reading %s\n", GRUB2_GRUB_CFG));
+        DEBUG (((gLinuxBootFlags & LINUX_BOOT_LOG_VERBOSE) == 0 ? DEBUG_VERBOSE : DEBUG_INFO,
+          "LNX: Reading %s\n", GRUB2_GRUB_CFG));
         Status = InternalProcessGrubCfg (GrubCfg);
       }
     }
@@ -747,22 +941,24 @@ InternalScanLoaderEntries (
   }
 
   if (!EFI_ERROR (Status)) {
-    gNamedLoaderEntries = OcFlexArrayInit (sizeof (NAMED_LOADER_ENTRY), (OC_FLEX_ARRAY_FREE_ITEM) InternalFreeNamedLoaderEntry);
-    if (gNamedLoaderEntries == NULL) {
+    gLoaderEntries = OcFlexArrayInit (sizeof (LOADER_ENTRY), (OC_FLEX_ARRAY_FREE_ITEM) InternalFreeLoaderEntry);
+    if (gLoaderEntries == NULL) {
       Status = EFI_OUT_OF_RESOURCES;
     } else {
-      Status = OcScanDirectory (EntriesDirectory, ProcessLoaderEntry, NULL);
+      Status = OcScanDirectory (EntriesDirectory, ProcessLoaderEntry, DirName);
     }
 
     if (!EFI_ERROR (Status)) {
-      Status = InternalConvertNamedLoaderEntriesToBootEntries (
+      ASSERT (gLoaderEntries->Count > 0);
+
+      Status = InternalConvertLoaderEntriesToBootEntries (
         RootDirectory,
         Entries,
         NumEntries
       );
     }
 
-    OcFlexArrayFree (&gNamedLoaderEntries);
+    OcFlexArrayFree (&gLoaderEntries);
   }
 
   InternalFreeGrubVars ();
@@ -779,64 +975,54 @@ InternalScanLoaderEntries (
   return Status;
 }
 
+STATIC
 EFI_STATUS
-ScanLoaderEntries (
-  IN   EFI_FILE_PROTOCOL        *RootDirectory,
+DoScanLoaderEntries (
+  IN   EFI_FILE_PROTOCOL        *Directory,
+  IN   CHAR16                   *DirName,
   OUT  OC_PICKER_ENTRY          **Entries,
   OUT  UINTN                    *NumEntries
   )
 {
   EFI_STATUS                      Status;
-  EFI_STATUS                      TempStatus;
-  EFI_FILE_PROTOCOL               *AdditionalScanDirectory;
 
-  Status = InternalScanLoaderEntries (RootDirectory, Entries, NumEntries);
-  if (!EFI_ERROR (Status)) {
-    return Status;
-  }
-  if (Status != EFI_NOT_FOUND) {
-    DEBUG ((DEBUG_WARN, "LNX: ScanLoaderEntries @root - %r\n", Status));
-  }
+  Status = ScanLoaderEntriesAtDirectory (Directory, DirName, Entries, NumEntries);
 
-  TempStatus = OcSafeFileOpen (RootDirectory, &AdditionalScanDirectory, ADDITIONAL_SCAN_DIR, EFI_FILE_MODE_READ, 0);
-  if (EFI_ERROR (TempStatus)) {
-    return Status;
-  }
-
-  Status = InternalScanLoaderEntries (AdditionalScanDirectory, Entries, NumEntries);
-  if (!EFI_ERROR (Status)) {
-    DEBUG ((DEBUG_INFO, "LNX: ScanLoaderEntries @%s - %r\n", ADDITIONAL_SCAN_DIR, Status));
-    return Status;
-  }
-  if (Status != EFI_NOT_FOUND) {
-    DEBUG ((DEBUG_WARN, "LNX: ScanLoaderEntries @%s - %r\n", ADDITIONAL_SCAN_DIR, Status));
-  }
+  DEBUG ((
+    (EFI_ERROR (Status) && Status != EFI_NOT_FOUND) ? DEBUG_WARN : DEBUG_INFO,
+    "LNX: ScanLoaderEntries %s - %r\n",
+    DirName,
+    Status
+    ));
+    
   return Status;
 }
 
-VOID
-InternalFreeNamedLoaderEntry (
-  NAMED_LOADER_ENTRY    *Entry
+EFI_STATUS
+InternalScanLoaderEntries (
+  IN   EFI_FILE_PROTOCOL        *RootDirectory,
+  OUT  OC_PICKER_ENTRY          **Entries,
+  OUT  UINTN                    *NumEntries
   )
 {
-  ASSERT (Entry != NULL);
-  ASSERT (Entry->Entry != NULL);
+  EFI_STATUS                    Status;
+  EFI_FILE_PROTOCOL             *AdditionalScanDirectory;
 
-  if (Entry != NULL) {
-    if (Entry->FileName != NULL) {
-      FreePool (Entry->FileName);
-      Entry->FileName = NULL;
-    }
-    
-    if (Entry->Entry != NULL) {
-      InternalFreeLoaderEntry (&(Entry->Entry));
+  Status = DoScanLoaderEntries (RootDirectory, ROOT_DIR, Entries, NumEntries);
+  if (EFI_ERROR (Status)) {
+    Status = OcSafeFileOpen (RootDirectory, &AdditionalScanDirectory, BOOT_DIR, EFI_FILE_MODE_READ, 0);
+    if (!EFI_ERROR (Status)) {
+      Status = DoScanLoaderEntries (AdditionalScanDirectory, BOOT_DIR, Entries, NumEntries);
+      AdditionalScanDirectory->Close (AdditionalScanDirectory);
     }
   }
+  
+  return Status;
 }
 
 STATIC
 EFI_STATUS
-DoConvertNamedLoaderEntriesToBootEntries (
+DoConvertLoaderEntriesToBootEntries (
   OUT  OC_PICKER_ENTRY          **Entries,
   OUT  UINTN                    *NumEntries
   )
@@ -844,7 +1030,6 @@ DoConvertNamedLoaderEntriesToBootEntries (
   EFI_STATUS                Status;
   UINTN                     Index;
   UINTN                     OptionsIndex;
-  NAMED_LOADER_ENTRY        *NamedEntry;
   LOADER_ENTRY              *Entry;
   OC_FLEX_ARRAY             *PickerEntries;
   OC_PICKER_ENTRY           *PickerEntry;
@@ -860,9 +1045,8 @@ DoConvertNamedLoaderEntriesToBootEntries (
   }
 
   Status = EFI_SUCCESS;
-  for (Index = 0; Index < gNamedLoaderEntries->Count; Index++) {
-    NamedEntry = OcFlexArrayItemAt (gNamedLoaderEntries, Index);
-    Entry = NamedEntry->Entry;
+  for (Index = 0; Index < gLoaderEntries->Count; Index++) {
+    Entry = OcFlexArrayItemAt (gLoaderEntries, Index);
 
     PickerEntry = OcFlexArrayAddItem (PickerEntries);
     if (PickerEntry == NULL) {
@@ -1032,16 +1216,16 @@ MakeAllUnique (
 {
   EFI_STATUS          Status;
   UINTN               Index;
-  NAMED_LOADER_ENTRY  *NamedEntry;
+  LOADER_ENTRY        *Entry;
 
   if (gPickerContext->HideAuxiliary) {
     return EFI_SUCCESS;
   }
 
-  for (Index = 0; Index < gNamedLoaderEntries->Count; Index++) {
-    NamedEntry = OcFlexArrayItemAt (gNamedLoaderEntries, Index);
+  for (Index = 0; Index < gLoaderEntries->Count; Index++) {
+    Entry = OcFlexArrayItemAt (gLoaderEntries, Index);
     
-    Status = MakeUnique (NamedEntry->Entry);
+    Status = MakeUnique (Entry);
     if (EFI_ERROR (Status)) {
       return Status;
     }
@@ -1058,8 +1242,8 @@ DisambiguateDuplicates (
 {
   UINTN               Index;
   UINTN               MatchIndex;
-  NAMED_LOADER_ENTRY  *NamedEntry;
-  NAMED_LOADER_ENTRY  *MatchEntry;
+  LOADER_ENTRY        *Entry;
+  LOADER_ENTRY        *MatchEntry;
 
   //
   // This check should not be strictly necessary, since there shouldn't
@@ -1069,23 +1253,23 @@ DisambiguateDuplicates (
     return EFI_SUCCESS;
   }
 
-  for (Index = 0; Index < gNamedLoaderEntries->Count - 1; Index++) {
-    NamedEntry = OcFlexArrayItemAt (gNamedLoaderEntries, Index);
+  for (Index = 0; Index < gLoaderEntries->Count - 1; Index++) {
+    Entry = OcFlexArrayItemAt (gLoaderEntries, Index);
 
-    if ((NamedEntry->DuplicateFlags & DUPLICATE_ID_SCANNED) != 0) {
+    if (Entry->DuplicateIdScanned) {
       continue;
     }
 
-    for (MatchIndex = Index + 1; MatchIndex < gNamedLoaderEntries->Count; MatchIndex++) {
-      MatchEntry = OcFlexArrayItemAt (gNamedLoaderEntries, MatchIndex);
+    for (MatchIndex = Index + 1; MatchIndex < gLoaderEntries->Count; MatchIndex++) {
+      MatchEntry = OcFlexArrayItemAt (gLoaderEntries, MatchIndex);
 
-      if ((MatchEntry->DuplicateFlags & DUPLICATE_ID_SCANNED) == 0) {
-        if (AsciiStrCmp (NamedEntry->Entry->OcId, MatchEntry->Entry->OcId) == 0) {
+      if (!MatchEntry->DuplicateIdScanned) {
+        if (AsciiStrCmp (Entry->OcId, MatchEntry->OcId) == 0) {
           //
           // Everything but the first id duplicate becomes auxiliary.
           //
-          MatchEntry->Entry->OcAuxiliary = TRUE;
-          MatchEntry->DuplicateFlags |= DUPLICATE_ID_SCANNED;
+          MatchEntry->OcAuxiliary = TRUE;
+          MatchEntry->DuplicateIdScanned = TRUE;
         }
       }
     }
@@ -1149,14 +1333,14 @@ ApplyDefaults (
 {
   EFI_STATUS          Status;
   UINTN               Index;
-  NAMED_LOADER_ENTRY  *NamedEntry;
+  LOADER_ENTRY        *Entry;
 
   Status = EFI_SUCCESS;
 
-  for (Index = 0; Index < gNamedLoaderEntries->Count; Index++) {
-    NamedEntry = OcFlexArrayItemAt (gNamedLoaderEntries, Index);
+  for (Index = 0; Index < gLoaderEntries->Count; Index++) {
+    Entry = OcFlexArrayItemAt (gLoaderEntries, Index);
 
-    Status = EntryApplyDefaults (RootDirectory, NamedEntry->Entry);
+    Status = EntryApplyDefaults (RootDirectory, Entry);
     if (EFI_ERROR (Status)) {
       break;
     }
@@ -1169,7 +1353,7 @@ ApplyDefaults (
 // Also use for loader entries generated by autodetect.
 //
 EFI_STATUS
-InternalConvertNamedLoaderEntriesToBootEntries (
+InternalConvertLoaderEntriesToBootEntries (
   IN   EFI_FILE_PROTOCOL        *RootDirectory,
   OUT  OC_PICKER_ENTRY          **Entries,
   OUT  UINTN                    *NumEntries
@@ -1180,12 +1364,12 @@ InternalConvertNamedLoaderEntriesToBootEntries (
   Status = EFI_SUCCESS;
 
   //
-  // Sort entries by filename descending.
+  // Sort entries by version descending.
   //
-  PerformQuickSort (gNamedLoaderEntries->Items, gNamedLoaderEntries->Count, gNamedLoaderEntries->ItemSize, OcReverseStringCompare);
+  PerformQuickSort (gLoaderEntries->Items, gLoaderEntries->Count, gLoaderEntries->ItemSize, InternalReverseVersionCompare);
 
   //
-  // Fill out any missing data, including from kernel file version string if needed.
+  // Autodetect good title and flavour, if needed.
   //
   Status = ApplyDefaults (RootDirectory);
 
@@ -1206,7 +1390,7 @@ InternalConvertNamedLoaderEntriesToBootEntries (
   }
 
   if (!EFI_ERROR (Status)) {
-    Status = DoConvertNamedLoaderEntriesToBootEntries (
+    Status = DoConvertLoaderEntriesToBootEntries (
       Entries,
       NumEntries
     );